requirements.txt
numpy==1.24.3
python-decouple==3.8
bison==0.1.3
numpy==1.24.3
pkg_resources==0.0.0
python-decouple==3.8
PyYAML==6.0
